Bot_update windows fix
git log is supposed to be called through git() rather than call() because
it adds ".bat" to the executable.  That was overlooked on the other waterfall
because chromium.git passes in git hashes directly, while these bots pass
in svn revision numbers, which triggers this code path.

BUG=261741

Review URL: https://codereview.chromium.org/209273002