Retry r259150: Determine language code and type of format for address.
BuildComponents() in address_ui.h should take the UI language and
determine the following two questions.

(1) Whether to show UI components based on local or latinized address
    format.

If the UI language does not match the default language for the country,
and latinized address format is available, then use the latinized
format. All countries with a latinized address format also have the
default language specified.

(2) Which language code should be used to format the address after the
    user provides it.

If the UI language is one of the languages supported by the country,
then use the UI language. If the country lists supported languages, none
of which match the UI language, then pick the first supported language.
If the country does not have supported languages (e.g. Antarctica), then
use the UI language for formatting. If using the latinized format and
the UI language is not one of the supported languages, then the
formatting language is [country's default language] + "-latn".

BUG=354954
Originally landed: http://crrev.com/259150
Reverted: http://crrev.com/259239

Review URL: https://codereview.chromium.org/198923010