Move some code from onCreate() to onStart() in Chromoting Android app
This reloads the system's accounts list whenever the app is started
(so if the user adds an account, then switches back to the app, the
change will be reflected).

This also reloads the host list, which will alleviate many of the
problems with stale OAuth tokens such as in bug 304719.

BUG=304719
NOTRY=true

Review URL: https://codereview.chromium.org/212673010