Fixing race conditions in ui::content::WindowSlider which could cause the overscroll overlay to never go away.
The race condition occurred because new naviagtion could happen before the callback from the animation associated with the old navigation was executed. This could get the OverscrollNavigationOverlay into a bad state.

The fix is to instantly complete the animation associated with a previous user action when new interaction occurs. Also, we now perform page navigation at the beginning of the transition animation, not in the end, which saves us 200ms in load time.

BUG=348717, 305448

Review URL: https://codereview.chromium.org/202183003