Fix resource leak code path around OpenManifestEntry.
Currently, there are some code paths which causes resource leak
in OpenManifestEntry (and its related methods).
There are mainly two cases:
1) If pnacl_options.translate() is true in original code,
open_cont is just leaked.
2) If stream_cc is created but not called, e.g. Plugin::StreamAsFile()
is failed somehow, or it is pNaCl related url, then the open_cont is
leaked.

It seems the root cause is PP_ComletionCallback/pp::CompletionCallback
usage. PP_CompletionCallback is designed for C, rather than C++,
so it does not manage the resources. pp::CompletionCallback is its
very simple wrapper. So, even when it is destructed without the invocation,
it just leakes the data.
Note that, WeakRef does not help such a case, because its resource management
fires only when the callback is invoked.
Probably, more generic wider solution would be minimize the usage
of these callbacks, such as the just interface to the PPAPI, and instead
base::Callback would be used internally in the code of chrome tree.

This CL just focuses on the OpenManifestEntry related code.

BUG=n/a
TEST=Ran trybots.

Review URL: https://codereview.chromium.org/211133003