[telemetry] Extract page summarization from buildbot to a new class
Some telemetry benchmark repeat a number of times in order to get a reliable
measurement. The test does not have to handle merging of these runs: summarizer
does it for you.

This used to be tangled up in the buildbot, and classes wanting to use the
summarization would subclass. the buildbot, inheriting all sorts of buildbot
oddities in the process. This extracts the summarizing behavior to a standalone
class.

BUG=340096
R=tonyg
NOTRY=True

Review URL: https://codereview.chromium.org/199633015