Block WebSocket handshake information transfer if necessary.
WebSocket handshake information is shown in the devtools, but it is sent
via the renderer process. Since it contains cookies, we should block it
when ChildProcessSecurityPolicyImpl::CanReadRawCookies returns false.

Since CanReadRawCookies is set only when the inspector is attached, we can
block other handshake information for the inspector as well as cookies when
it is unset, for simplicity and performance. In fact, this CL avoids sending
NotifyStartOpeningHandshake and NotifyFinishOpeningHandshake IPCs
in such a case.

BUG=310405

Review URL: https://codereview.chromium.org/212523002