Add Autofill preview support for <select> input fields
Since changing the text in display of <select> will inevitably change the selectedIndex of the options, a workaround is introduced for this case. In the workaround,
it saves a copy of the original selected option and recovers this value once autofill preview is done. 

BUG=58719
R=dhollowa@chromium.org, isherman@chromium.org, jhawkins@chromium.org

Review URL: https://codereview.chromium.org/136793007