sync: Add NonBlockingTypeProcesssor and SyncCore
Introduces the model-thread sibling of the NonBlockingTypeProcessorCore
that was added in r258390.  Also adds SyncCore and SyncCoreProxy, the
classes that will be used to connect the two.

The SyncCore lives on the sync thread and carries out requests sent to
it by the SyncCoreProxy.  The SyncCoreProxy is a thread-safe wrapper
around the SyncCore that can be easily copied and whose methods can be
called from any thread.

The NonBlockingTypeProcessor is instantiated on the same thread as the
data to be synced.  It connects to a NonBlockingTypeProcessorCore on the
sync thread by sending a request through a SyncCoreProxy.  Keeping data
in sync will be a collaborative effort involving both the
NonBlockingTypeProcessor and NonBlockingTypeProcesssorCore, though none
of this functionality has been implemented yet.

As of this CL, none these classes are instantiated outside of tests.
This CL should have no effect on current sync behavior.

BUG=351005

Review URL: https://codereview.chromium.org/208893004