Remove pattern from GN and clang format deps entries.
The pattern facility for checking if a download is out-of-date is broken in a number of circumstances, which will result in out-of-date deps. This change will force a SHA check every time, but that should be relatively fast compared to the hooks run, and will eliminate some flakiness reported by some engineers.

BUG=
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/214663006