Add a whole-app-list unit test.
This establishes a test harness for doing tests that touch on multiple
parts of the app launcher UI, or that need coverage for the various app
launcher configurations: Ash or Desktop; regular or experimental
"landscape" mode.

It currently just tests displaying and closing the app launcher in its
default state with a flat (no folders), 2- or 3-page model.

BUG=169114
TEST=Added unit tests AppListViewTest{Aura,Desktop}.*

Review URL: https://codereview.chromium.org/210083009