Split InfoBarService core code into InfoBarManager
InfoBarService is now responsible for associating an InfoBarManager to a Tab
and handling the interactions with the rest of chrome (notifications, navigation
events).

InfoBarManager has the core logic for infobar management. It owns the list of infobars.

Eventually InfoBarManager will no longer rely on content/, but for now it still
has a reference to a WebContents instance. This reference was kept for now to 
minimize the size of the change, but will be removed in a future CL.

BUG=354379
TBR=jochen

Review URL: https://codereview.chromium.org/211273007