cc: Avoid prefer smoothness mode when scrolling with scroll handler
Avoid going into prefer smoothness mode when the scrolled layer has a
scroll handler. The presence of a scroll handler suggests that the main
thread wants to animate in response to scrolling. Avoiding smoothness
mode gives it a chance to do that while still having our deadline as a
fallback.
    
BUG=347366

Review URL: https://codereview.chromium.org/206523003