linux_aura: Port GtkKeybindingsHandler to Aura.
This is a quick hack to move the current GtkKeybindingsHandler code into
aura builds. This is necessary so that Linux users who have set custom
keybindings can use them in edit fields. This only works in the content
area; further patches are needed for this to work in views.

This redoes the interface so that instead of being hard coded into
content, we have a general interface set in ui/wm/. This should allow
views to also use this code in ui/views/, as well as its current usage
in content/.

BUG=319437, 358297
R=piman@chromium.org, sadrul@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/213283004