cc: Add tiling eviction tile iterator.
This patch adds a naive version of the eviction iterator for a tiling.
This sets the interface for the class, so that work can continue on
handing eviction from the layer and tile manager, as well as being able
to independently change the implementation for something more optimal.

R=enne

Review URL: https://codereview.chromium.org/217033004