Fix bitmap fetcher race condition.
There is no need to re-fetch a bitmap if we are already fetching it.
The code can get into this race condition when processing synced notifications,
with one notification being loaded from sync stored data, and a newer version
of the same notification coming in from the wire (from the sync server).

BUG=358240

Review URL: https://codereview.chromium.org/218633005