metrics: Add UpdateEngine.* metrics.
We've rewritten the metrics reported by the ChromeOS update_engine.
It's done in a way so the new metrics (UpdateEngine.*) will live
alongside the with the existing metrics (Installer.*). We plan to
eventually remove the old metrics.

This CL adds miscellaneous metrics that don't naturally fit in any of
the UpdateEngine.{Check,Attempt,Successful} namespaces.

BUG=chromium:355745
TEST=Validated the xml, pretty print succeeded.
R=avakulenko@chromium.org, deymo@chromium.org, isherman@chromium.org, jar@chromium.org, sosa@chromium.org

Review URL: https://codereview.chromium.org/221243004