CHECK that ui::Windows being destroyed have been hidden if necessary
There have been a bunch of crashes due to the GestureRecognizer targeting
touches to windows that have been destroyed. When a window is hidden 
(in WindowEventDispatcher::OnWindowHidden), all references to that window 
are removed from the GestureRecognizer.

It looks like some windows with touches targeted to them are being 
destroyed without being hidden first. This change CHECKs that this
never occurs.

BUG=342040

Review URL: https://codereview.chromium.org/224063002