Update KeystoneRegistration.framework to 1.2.0.7407.
The big change in this version is that an installation attempt initiated by
the application can show up as ondemand on the server. This will allow us to
provide throttled Mac downloads (such as a "5% push") with correct behavior.

A previous registration framework, 1.2.0.6782, intended to fix this bug was
backed out because it was not compatible with older installed Keystones. This
version contains a fix for that problem.

Previous commit: r245361, https://codereview.chromium.org/141193002
Backed out: r245603, https://codereview.chromium.org/141583005

BUG=35495

Review URL: https://codereview.chromium.org/212703006