android_webview: Update top-level makefile for multiarch.
Add multiarch support (64+32 bit parallel builds) to the top-level
Android.mk: include two sets of makefiles if needed, one for each of the
two target architectures.

This also cleans up the general structure: nobody builds the regular
Android Chromium inside of an Android tree any more, so it's no longer
necessary to handle the case where this Android.mk is being parsed by
the build system but it's *not* intended to be building WebView.

BUG=358141
R=benm@chromium.org

Review URL: https://codereview.chromium.org/225593004