Store the id of a contributing motion event in GestureEventData
The unified GR needs to be able to construct ui::GestureEvents from
GestureEventData objects. Currently no information on
contributing touch ids is available in GestureEventData.

This information is used for figuring out which window to target 
a gesture to - the gesture is targeted to the window which the 
contributing touch is locked to.

BUG=332418

Review URL: https://codereview.chromium.org/212663010