Add extensions_unittests target for unit tests in //extensions.
This introduces a minimal test target, "extensions_unittests", for
building unit tests in //extensions. It starts with files from
extensions/common first, so it is easier to get the dependencies right.

This also list the all the entries that would not introduce more
"chrome" dependencies on any extensions targets.

When running this target you will see that it does not PASS everything
yet:

8 tests failed:
    ExtensionURLPatternTest.ConvertToExplicitSchemes
    ExtensionURLPatternTest.GetAsString
    ExtensionURLPatternTest.Match13
    ExtensionURLPatternTest.Match19
    ExtensionURLPatternTest.Match9
    FileUtilTest.ExtensionResourceURLToFilePath
    FileUtilTest.ExtensionURLToRelativeFilePath
9 tests crashed:
    APIPermissionSetTest.CreateDifference
    APIPermissionSetTest.CreateIntersection
    APIPermissionSetTest.CreateUnion
    APIPermissionSetTest.General
    APIPermissionSetTest.IPC
    APIPermissionSetTest.ImplicitPermissions
    ManifestHandlerTest.DependentHandlers
    ManifestHandlerTest.FailingHandlers
    ManifestHandlerTest.Validate

Since we are not running this target on the bots yet, it should be fine,
and we can iterate from this point and fix these failures and crashes later.

BUG=348066
TEST=extensions_unittests
R=yoz@chromium.org,jamescook@chromium.org,jam@chromium.org

Review URL: https://codereview.chromium.org/224993004