simple cache: testing: AddDelay() between evictable entry sets
We switched to very coarse-grained timestamps in the backend. This requires
adding synthetic delays in tests to make different timestamps for different
subsets of entries. We added one delay in the test, but forgot about the other.

This change separates the last two entries from the rest with a delay. With the
delay the current eviction algorithm (i.e. sort by timestamp, evict down to the
low watermark), should not evict these two last entries.

BUG=234776
TBR=gavinp@chromium.org

Review URL: https://codereview.chromium.org/225833002