Password bubble: Swap out views for the ItemView's various states.
Currently, we set up a GridLayout in MannagePasswordItemView, populate
it with a Label and a Link, and swap out the values of those two Views.
This a) is complex, and b) doesn't allow us to correctly align the data
when switching between two-column (Undo) and three-column (Management)
views.

This patch turns ManagePasswordItemView into a container, holding a
single child view, laid out as a FillLayout. Three child Views represent
the various states the view can find itself in; views are created and
destroyed as appropriate for the view's current state.

BUG=261628

Review URL: https://codereview.chromium.org/225753002