Update settings UI when ownership information changes.
Ownership assignment of a device can take up to tens of seconds on some
ChromeOS devices and if the settings UI is opened before that it will
show wrong information that the current user is not the owner even if he
is. This change fixes this.

BUG=349110

Review URL: https://codereview.chromium.org/223553004