Improving the user transition to add special cases for maximized windows and make the transition "more fluid"
There were user complaints with the old animation which said that the animation looked janky. This was in part because various transitions (e.g. maximized->maximized) were showing the background animating in the middle of the animation which looked like "flashing") and because there are some real causes where the CPU gets 100% hogged and the UI thread has not time processing the animation.

This is addressing the first part: Improving the animation by removing the "flash" in the middle and instead transitioning the desktop cleanly from one state to another.

The shelf animation stays the same, but the windows will cross dissolve. When switching from maximize to maximize, the desktop will not shine through and when switching from / or to maximize, the desktop background will only show the desktop background from the user which is not maximized. This way the intermediate frames do not show "some state".

BUG=357852
TEST=unittest & visual checking

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=261944

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=262024

Review URL: https://codereview.chromium.org/223823004