sync: Re-implement getAllNodes WebUI function
Removes the existing implementation of getAllNodes.  This was the last
function based on the old "makeSyncFunction" and "JsMessageHandler"
interface, so its removal leaves behind quite a bit of dead code.  This
CL removes some of it.  The rest will be removed in a future CL.

Replaces it with some infrastructure intended to be more compatible with
the future of sync, where each type is more independent.  Requests to
getAllNodes are routed through a DirectoryTypeDebugInfoEmitter in the
ModelTypeRegistry.  A NonBlockingTypeDebugInfoEmitter will be
implemented in a future CL.

The new system also intended to support "streaming" results.  Rather
than waiting for all types to return their nodes, the system could be
modified to allow some types to get back to the JavaScript layer sooner
than others so it can display results earlier.  However, we do not
currently take advantage of this functionality.

The return type of getAllNodes is now an array of per-type objects,
rather than one big list of nodes.  This, too, should help us implement
streaming support in the future.  For now, most of the JavaScript
callbacks will convert the data back to the old format before operating
on it.

BUG=328606

Review URL: https://codereview.chromium.org/224563004