V4L2: Connect antibanding/anti-flicker where supported
Some Linux UVC devices implement the V4L2 User Controls
API [1]. For them, set the antibanding/anti-flicker/power
frequency rejection to the appropriate frequency (50/60Hz).
This settings sinks all the way into the WebCam DSP so
it's hard to know for sure what is being done.

The current country is extracted from the system timezone
using CountryCodeForCurrentTimezone(). From its doc:
"Checks the system timezone and turns it into a two-character
ASCII country code. This may fail (for example, it will
always fail on Android), in which case it will return
an empty string." [2]

TEST: In Linux, the anti-flicker API can be exercised
using f.i. `guvcview` application. Find a spot with 
convenient flickering, f.i. a wall with a fluorescent 
light falling on it, specially with dim general 
illumination conditions, and open this program, then
connect the appropriate antibandings to witness the 
flicker removal.

The removal effect can be seen in the recording [3].

[1] http://www.linuxtv.org/downloads/legacy/video4linux/API/V4L2_API/spec-single/v4l2.html#control
[2] https://code.google.com/p/chromium/codesearch#chromium/src/base/i18n/timezone.h&q=CountryCodeForCurrentTimezone&sq=package:chromium&type=cs&l=17
[3] https://drive.google.com/a/google.com/file/d/0BwgFm5xOT0yCTHFKejBsQVRGczQ/edit?usp=sharing

BUG=357599

TBR=tommi@chromium.org

Review URL: https://codereview.chromium.org/216263007