Add a vulcanized version of polymer to third_party
This adds a vulcanized version of polymer to third_party so that library
polymer elements can be used by extensions/apps. This also adds a python
script for updating this version of polymer and a PRESUBMIT check to make
sure it is run at the right time.

This needs to be done because library polymer elements aren't CSP compliant
however Chrome extensions/apps have CSP enforced. More specifically this is being added for the PDF extension. See the associated bug.

This also adds the missing polymer-body.html.

BUG=359333,303491

Review URL: https://codereview.chromium.org/227423002