[Android] Reland "Provide unhandled tap event notifications"
Allow notifications of unhandled taps via the GestureStateListener.  This change
makes GestureTap events blocking, but there are several use-cases for which this
will be necessary, e.g., WebView and contextual search.

This was landed in r26143 and reverted in r261470 due to to a failure in
TouchEditableImplAuraTest.TouchCursorInTextfieldTest.  The test has been fixed.

Review URL: https://codereview.chromium.org/218633008