Remove evaluateJavaScript and getRenderCoordinates from ContentView.
They were just forwarding on to CVC so just call it directly.

DOMUtils is left in a bit of awkward place. Because it has some
functions that relate to clicking, it still needs ContentView in some
places.

Additionally, we could further simplify some of the tests to not refer
to ContentView but that requires changing TestCallbackHelperContainer's
constructor which is used in a lot of places so leaving for a separate
change.

BUG=360664

Review URL: https://codereview.chromium.org/227603010