cc: Make RasterWorkerPool an abstract class.
This makes ownership more clear and avoids the need to have the
base class call virtual functions that need to be implemented by
derived classes. Also removes the need for the base class to
implement the WorkerPoolTaskClient interface.

No changes in behavior. Might have a small positive impact on
performance as it removes an unnecessary level of callback
redirection from the "raster finished" tasks.

BUG=269841

Review URL: https://codereview.chromium.org/221953003