Make ContentExtractor more automatable
This adds three options to ContentExtractor

--disable-dns disables dns. I figured that leaving the default at dns
enabled would be good since that's probably what you want when manually
running the content extractor (and it's easier for scripts/automation to
pass extra flags than for humans to).

--output-file writes output to a file instead of stdout. This makes it
so automation doesn't have to parse the relevant output from all the
other logging and test stuff that is printed.

--output-binary writes a serialized protobuf instead of human-readable
output.

TBR=jochen@chromium.org

Review URL: https://codereview.chromium.org/214603009