Prevent bookmark apps from triggering a sync loop.
This CL makes bookmark apps sync a dummy version as their actual version
is based on creation time and makes no sense to sync.

This fixes an issue where the version number of bookmark apps wouldn't
match down-synced data so that when an up-sync was triggered, the version
number on the server was always overwritten (possibly triggering another down
and up sync on another machine).

BUG=361003

Review URL: https://codereview.chromium.org/226963003