Flush after LTHI::ReclaimResources when hidden.
In ReclaimResources we return resources from the parent, which usually end up
deleted when we're invisible (since our allocation is supposed to be 0). In that
case we want to make sure that we flush the context so that the textures are
actually deleted in the GPU process (and free up memory for the rest of the
system).

BUG=349553

Review URL: https://codereview.chromium.org/229893003