Show a notification when an ephemeral app consumes excessive disk space
This patch introduces monitoring of the storage usage of extensions
and apps that are granted unlimited storage. A notification is shown
when persistent storage usage exceeds a threshold.

Monitoring is implemented for all extensions and apps, but initially
only enabled for ephemeral apps. Ephemeral apps have a lower threshold.

BUG=347801
TEST=browser_tests (ExtensionStorageMonitor.*)

Review URL: https://codereview.chromium.org/221933013