Change the Pica load benchmark to listen for the polymer-ready event
In the last data update, 005, Pica changed its boot-up sequence to no
longer depend on the old WebComponentsReady event, but the page set
was still using that to decide that the load was finished.

This patch switches the Pica page set to wait for 'polymer-ready' instead,
which should be fired just before the first paint.

It also includes a new WPR archive that cuts the payload size by 1/3.

BUG=361367
TBR=tonyg@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/232203002