Ensure FS is exited for all not-in-same-page navigations.
This is a revisit of https://codereview.chromium.org/218093002.  In that
change, I assumed that if was_same_page == true, then the fullscreen
exit should not be forced (e.g., a navigation that is just a hash tag
change in the URL).  However, this flag turns out to be unreliable:
There are cases (see bug) where going back/forward between two different
pages on the same site will commit a navigation with
was_same_page == true.

This change instead asks the NavigationController to compare the new URL
versus the last-committed entry's URL to decide.

BUG=356951

Review URL: https://codereview.chromium.org/230043002