Enable Error Console only for Unpacked extensions
Having it enabled for all extensions leads to too much noise. Let's only have
it for unpacked extensions for now.

We have the backend implementation to turn it on or off for only particular
extensions (and error severity), but we don't have the UI for it yet. In the
interim, let's do this.

BUG=360459
TBR=finnur@chromium.org (for small changes to extension_settings_handler.cc)

Review URL: https://codereview.chromium.org/231523002